requirements.txt
numpy==1.18.1
pandas==0.25.3
pillow==5.1.0
h5py==2.10.0
tensorflow==2.0.0
keras==2.3.1
